requirements.txt
requests
beautifulsoup4
openai
selenium
python-dotenv
webdriver_manager